feat: add schema spec and collection helpers
